msm: kgsl: Set HIGH_PRI to KGSL workqueue and bound work items
As we are dealing with GPU processes, these should be given more prio than normal work pools. Thus, let's give the workqueue a WQ_HIGHPRI flag. Also since the work items being queued to this workqueues are likely to be triggered a lot of times, let's bound them according to their importance in the code structure for them to benefit with increased cache locality as well as more perf as for the processes that are of more importance than others.

Signed-off-by: ederekun <sedrickvince@gmail.com>