sched/core: Make HRTICK run properly at suspend
Use timekeeping_suspended (which is also a warning issued by ktime_get if used at suspend) to decide when to use an alternate get_time function that is allowed at suspend and is still being updated by timekeeping at the expense of inaccuracy at and caused by the transition from an active state to suspension. With this change, we must switch to boottime base clock in order to take into consideration the offset caused by suspend time even during active periods to keep consistency as much as possible.

Signed-off-by: ederekun <sedrickvince@gmail.com>