fork: Call dynamic_stune as soon as possible
The crucial parameter affects cpu placement which happens before cgroup proc write which makes the function ineffective. Thus, let's instead trigger it at fork together with the recently moved task_belongs_to_st function wherein we could now determine the st cgroup of the parent proc/task that have forked the process which makes it possible to check if it is in top-app st cgroup.

With this change, let's rename the dynamic stune structure to FORK to represent the function caller.

Signed-off-by: ederekun <sedrickvince@gmail.com>